refactor: merge Eyes into Body tab, promote POST to top-level page
Eyes tab content (eye prescriptions CRUD) now lives under the Body tab.
POST moved out of MeatSpace into its own top-level route at /post.
Replaced refreshKey re-fetch pattern with reactive local state updates.
