+++ client/src/components/meatspace/constants.js
@@ -1,17 +1,14 @@
 import {
-  Skull,
   Beer,
   Calendar,
   Clock,
   Dna,
-  Eye,
   HeartPulse,
   ClipboardList,
   Activity,
   Scale,
   Stethoscope,
-  Upload,
-  Zap
+  Upload
 } from 'lucide-react';
 
 export const TABS = [
@@ -21,12 +18,10 @@ export const TABS = [
   { id: 'blood', label: 'Blood', icon: HeartPulse },
   { id: 'body', label: 'Body', icon: Scale },
   { id: 'calendar', label: 'Calendar', icon: Calendar },
-  { id: 'eyes', label: 'Eyes', icon: Eye },
   { id: 'genome', label: 'Genome', icon: Dna },
   { id: 'health', label: 'Health', icon: Stethoscope },
   { id: 'import', label: 'Import', icon: Upload },
   { id: 'lifestyle', label: 'Lifestyle', icon: ClipboardList },
-  { id: 'post', label: 'POST', icon: Zap }
 ];
 
 // Lifestyle adjustment table for death clock
