+++ client/src/pages/MeatSpace.jsx
@@ -9,12 +9,10 @@ import AlcoholTab from '../components/meatspace/tabs/AlcoholTab';
 import BloodTab from '../components/meatspace/tabs/BloodTab';
 import BodyTab from '../components/meatspace/tabs/BodyTab';
 import CalendarTab from '../components/meatspace/tabs/CalendarTab';
-import EyesTab from '../components/meatspace/tabs/EyesTab';
 import GenomeTab from '../components/meatspace/tabs/GenomeTab';
 import HealthTab from '../components/meatspace/tabs/HealthTab';
 import ImportTab from '../components/meatspace/tabs/ImportTab';
 import LifestyleTab from '../components/meatspace/tabs/LifestyleTab';
-import PostTab from '../components/meatspace/tabs/PostTab';
 
 export default function MeatSpace() {
   const { tab } = useParams();
@@ -39,8 +37,6 @@ export default function MeatSpace() {
         return <BodyTab />;
       case 'calendar':
         return <CalendarTab />;
-      case 'eyes':
-        return <EyesTab />;
       case 'genome':
         return <GenomeTab />;
       case 'health':
@@ -49,8 +45,6 @@ export default function MeatSpace() {
         return <ImportTab />;
       case 'lifestyle':
         return <LifestyleTab />;
-      case 'post':
-        return <PostTab />;
       default:
         return <OverviewTab />;
     }
